README.md: Make info box sound more official (#25)
- [x] This contribution adheres to
[CONTRIBUTING.md](https://github.com/ni/nisync-python/blob/main/CONTRIBUTING.md).

### What does this Pull Request accomplish?

Update the info box text.

### Why should this Pull Request be merged?

"a Python API for..." implies that there is more than one.

We don't need to include the GitHub link on the first line of the
readme. Users who view the readme on GitHub are already here, and users
who view the readme on PyPI can click the "repository" link on the left
sidebar.

### What testing has been done?

N/A

Signed-off-by: Brad Keryan <brad.keryan@ni.com>